Added a fake data maker. This basically makes a grid of pixels with a gaussian feature at the given x,y coordinate.
This could be used as training data. Basically, for a gaussian centered at a particular point, you should be able to figure out that point! Ideally, you should not be "smart" about it and do a.. say, max-finding algorithm, which would totally work for this, but that's not what we want! Maybe instead, you should have your algorithm spit out the coorinate that is... 20 pixels to the RIGHT of the gaussian (which means you'll have to shift the coords xx_yy -> xx+20_yy or something).

A c++ program "testdata" makes the data for a single data run.
The script "run.sh" runs this program multiple times and makes a bunch of data runs, labeled by filename.